requirements.txt
uvicorn
watchfiles
python-dotenv
sqlalchemy
fastapi
fastmcp
openai
aiosqlite
aiomysql
psycopg2-binary
pydantic
requests
qdrant-client
pydantic-settings

pytest
python-dotenv
openai
httpx
greenlet
gspread

google-auth
mysql-connector-python
pytest-asyncio
pytest-mock